move some test utilities like timeline out of Hub tests
